Migrated dummy years data into its own file for better organization
