+++ advisor-ui/sampleYearData.js
@@ -1,0 +1,145 @@
+const years = [
+  {
+    number: 2,
+    semesters: [
+      {
+        number: 1,
+        classes: [
+          {
+            title: "Calculus 3",
+            units: 8,
+          },
+          {
+            title: "Biology",
+            units: 8,
+          },
+          {
+            title: "Chemistry",
+            units: 8,
+          },
+          {
+            title: "AP Literature",
+            units: 8,
+          },
+        ],
+      },
+      {
+        number: 2,
+        classes: [
+          {
+            title: "English 3",
+            units: 8,
+          },
+          {
+            title: "Cooking",
+            units: 8,
+          },
+          {
+            title: "Geometry",
+            units: 8,
+          },
+          {
+            title: "Physics",
+            units: 8,
+          },
+        ],
+      },
+    ],
+  },
+  {
+    number: 3,
+    semesters: [
+      {
+        number: 1,
+        classes: [
+          {
+            title: "Calculus 3",
+            units: 8,
+          },
+          {
+            title: "Biology",
+            units: 8,
+          },
+          {
+            title: "Chemistry",
+            units: 8,
+          },
+          {
+            title: "AP Literature",
+            units: 8,
+          },
+        ],
+      },
+      {
+        number: 2,
+        classes: [
+          {
+            title: "English 3",
+            units: 8,
+          },
+          {
+            title: "Cooking",
+            units: 8,
+          },
+          {
+            title: "Geometry",
+            units: 8,
+          },
+          {
+            title: "Physics",
+            units: 8,
+          },
+        ],
+      },
+    ],
+  },
+  {
+    number: 4,
+    semesters: [
+      {
+        number: 1,
+        classes: [
+          {
+            title: "Calculus 3",
+            units: 8,
+          },
+          {
+            title: "Biology",
+            units: 8,
+          },
+          {
+            title: "Chemistry",
+            units: 8,
+          },
+          {
+            title: "AP Literature",
+            units: 8,
+          },
+        ],
+      },
+      {
+        number: 2,
+        classes: [
+          {
+            title: "English 3",
+            units: 8,
+          },
+          {
+            title: "Cooking",
+            units: 8,
+          },
+          {
+            title: "Geometry",
+            units: 8,
+          },
+          {
+            title: "Physics",
+            units: 8,
+          },
+        ],
+      },
+    ],
+  },
+];
+
+export { years };
