+++ advisor-ui/src/pages/ScheduleTool/ScheduleTool.jsx
@@ -5,7 +5,7 @@ import { Link } from "react-router-dom";
 import "./ScheduleTool.css";
 import Constraints from "../../components/Constraints/Constraints";
 import ScheduleDetails from "../../components/ScheduleDetails/ScheduleDetails";
-
+import { years } from "../../../sampleYearData";
 export default function ScheduleTool() {
   const [constraints, setConstraints] = useState([]);
   const [schedule, setSchedule] = useState([]);
@@ -39,150 +39,6 @@ export default function ScheduleTool() {
   const generateNewSchedule = () => {
     setSchedule(years);
   };
-
-  const years = [
-    {
-      number: 2,
-      semesters: [
-        {
-          number: 1,
-          classes: [
-            {
-              title: "Calculus 3",
-              units: 8,
-            },
-            {
-              title: "Biology",
-              units: 8,
-            },
-            {
-              title: "Chemistry",
-              units: 8,
-            },
-            {
-              title: "AP Literature",
-              units: 8,
-            },
-          ],
-        },
-        {
-          number: 2,
-          classes: [
-            {
-              title: "English 3",
-              units: 8,
-            },
-            {
-              title: "Cooking",
-              units: 8,
-            },
-            {
-              title: "Geometry",
-              units: 8,
-            },
-            {
-              title: "Physics",
-              units: 8,
-            },
-          ],
-        },
-      ],
-    },
-    {
-      number: 3,
-      semesters: [
-        {
-          number: 1,
-          classes: [
-            {
-              title: "Calculus 3",
-              units: 8,
-            },
-            {
-              title: "Biology",
-              units: 8,
-            },
-            {
-              title: "Chemistry",
-              units: 8,
-            },
-            {
-              title: "AP Literature",
-              units: 8,
-            },
-          ],
-        },
-        {
-          number: 2,
-          classes: [
-            {
-              title: "English 3",
-              units: 8,
-            },
-            {
-              title: "Cooking",
-              units: 8,
-            },
-            {
-              title: "Geometry",
-              units: 8,
-            },
-            {
-              title: "Physics",
-              units: 8,
-            },
-          ],
-        },
-      ],
-    },
-    {
-      number: 4,
-      semesters: [
-        {
-          number: 1,
-          classes: [
-            {
-              title: "Calculus 3",
-              units: 8,
-            },
-            {
-              title: "Biology",
-              units: 8,
-            },
-            {
-              title: "Chemistry",
-              units: 8,
-            },
-            {
-              title: "AP Literature",
-              units: 8,
-            },
-          ],
-        },
-        {
-          number: 2,
-          classes: [
-            {
-              title: "English 3",
-              units: 8,
-            },
-            {
-              title: "Cooking",
-              units: 8,
-            },
-            {
-              title: "Geometry",
-              units: 8,
-            },
-            {
-              title: "Physics",
-              units: 8,
-            },
-          ],
-        },
-      ],
-    },
-  ];
 
   return (
     <div className="scheduletool">
